replaced long with int
git-svn-id: https://btanks.svn.sourceforge.net/svnroot/btanks/trunk@408 5c0bf991-d514-0410-a65f-ce4832e571f3

+++ tmx/layer.cpp
@@ -1,6 +1,6 @@
 #include "layer.h"
 #include <assert.h>
 
-Layer::Layer(const long w, const long h, const mrt::Chunk & data, const int impassability) : impassability(impassability), _w(w), _h(h){
+Layer::Layer(const int w, const int h, const mrt::Chunk & data, const int impassability) : impassability(impassability), _w(w), _h(h){
 	_data = data;
 }
